+++ crates/graph/src/range/exec/mem_ops/concat.rs
@@ -30,17 +30,19 @@ impl RangeConcat<Concrete, RangeConcrete<Concrete>> for RangeDyn<Concrete> {
                         let mut bytes = [0x00; 32];
                         bytes[0] = *v;
                         let v = Elem::from(Concrete::Bytes(1, H256::from(bytes)));
-                        (idx, (v, self.op_num + i))
+                        (idx, (v, self.op_num + i + 1))
                     })
                     .collect::<BTreeMap<_, _>>();
                 existing.extend(new);
                 Some(Elem::ConcreteDyn(RangeDyn::new_w_op_nums(
-                    Elem::from(Concrete::from(U256::from(val.len()))),
+                    (*self.len).clone() + Elem::from(Concrete::from(U256::from(val.len()))),
                     existing,
                     other.loc,
                 )))
             }
-            (Concrete::String(val), inner) if inner.is_none() || inner.unwrap().is_string() => {
+            (Concrete::String(val), inner)
+                if inner.is_none() || inner.unwrap().is_string() || inner.unwrap().is_bytes() =>
+            {
                 let last = self.len.clone();
                 let mut existing = self.val.clone();
                 let new = val
@@ -52,17 +54,20 @@ impl RangeConcat<Concrete, RangeConcrete<Concrete>> for RangeDyn<Concrete> {
                         let mut bytes = [0x00; 32];
                         v.encode_utf8(&mut bytes[..]);
                         let v = Elem::from(Concrete::Bytes(1, H256::from(bytes)));
-                        (idx, (v, self.op_num + i))
+                        (idx, (v, self.op_num + i + 1))
                     })
                     .collect::<BTreeMap<_, _>>();
                 existing.extend(new);
                 Some(Elem::ConcreteDyn(RangeDyn::new_w_op_nums(
-                    Elem::from(Concrete::from(U256::from(val.len()))),
+                    (*self.len).clone() + Elem::from(Concrete::from(U256::from(val.len()))),
                     existing,
                     other.loc,
                 )))
             }
-            _e => None,
+            e => {
+                debug_assert!(false, "was not concattable type: {e:#?}");
+                None
+            }
         }
     }
 }
@@ -80,17 +85,13 @@ impl RangeConcat<Concrete, RangeDyn<Concrete>> for RangeDyn<Concrete> {
                     .clone()
                     .into_iter()
                     .enumerate()
-                    .map(|(i, (key, (v, _op)))| (key + *last.clone(), (v, self.op_num + i)))
+                    .map(|(i, (key, (v, _op)))| (key + *last.clone(), (v, self.op_num + i + 1)))
                     .collect::<BTreeMap<_, _>>();
 
                 existing.extend(other_vals);
 
                 Some(Elem::ConcreteDyn(RangeDyn::new_w_op_nums(
-                    *self.len.clone()
-                        + *other
-                            .len
-                            .clone()
-                            .max(Box::new(Elem::from(Concrete::from(U256::from(1))))),
+                    *self.len.clone() + *other.len.clone(),
                     existing,
                     other.loc,
                 )))
@@ -145,5 +146,153 @@ pub fn exec_concat(
         Some(candidates.remove(candidates.len() - 1))
     } else {
         Some(candidates.remove(0))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::DummyGraph;
+    use solang_parser::pt::Loc;
+    use pretty_assertions::{assert_eq, assert_ne};
+
+    #[test]
+    fn concrete_concrete_bytes() {
+        let x = RangeConcrete::new(
+            Concrete::from(vec![b'h', b'e', b'l', b'l', b'o']),
+            Loc::Implicit,
+        );
+        let y = RangeConcrete::new(
+            Concrete::from(vec![b'w', b'o', b'r', b'l', b'd']),
+            Loc::Implicit,
+        );
+        let expected = Concrete::from(vec![
+            b'h', b'e', b'l', b'l', b'o', b'w', b'o', b'r', b'l', b'd',
+        ]);
+        let result = x.range_concat(&y).unwrap().maybe_concrete_value().unwrap();
+        assert_eq!(result.val, expected);
+    }
+
+    #[test]
+    fn concrete_concrete_bytes_str_fail() {
+        let x = RangeConcrete::new(
+            Concrete::from(vec![b'h', b'e', b'l', b'l', b'o']),
+            Loc::Implicit,
+        );
+        let y = RangeConcrete::new(Concrete::from("world"), Loc::Implicit);
+        assert!(x.range_concat(&y).is_none());
+    }
+
+    #[test]
+    fn concrete_concrete_bytes_none() {
+        let x = RangeConcrete::new(
+            Concrete::from(vec![b'h', b'e', b'l', b'l', b'o']),
+            Loc::Implicit,
+        );
+        let y = RangeConcrete::new(Concrete::DynBytes(vec![]), Loc::Implicit);
+        let result = x.range_concat(&y).unwrap().maybe_concrete_value().unwrap();
+        assert_eq!(result.val, x.val);
+    }
+
+    #[test]
+    fn concrete_concrete_str() {
+        let x = RangeConcrete::new(Concrete::from("hello"), Loc::Implicit);
+        let y = RangeConcrete::new(Concrete::from("world"), Loc::Implicit);
+        let expected = Concrete::from("helloworld");
+        let result = x.range_concat(&y).unwrap().maybe_concrete_value().unwrap();
+        assert_eq!(result.val, expected);
+    }
+
+    #[test]
+    fn concrete_concrete_str_bytes_fail() {
+        let x = RangeConcrete::new(Concrete::from("world"), Loc::Implicit);
+        let y = RangeConcrete::new(
+            Concrete::from(vec![b'h', b'e', b'l', b'l', b'o']),
+            Loc::Implicit,
+        );
+        assert!(x.range_concat(&y).is_none());
+    }
+
+    #[test]
+    fn concrete_concrete_str_none() {
+        let x = RangeConcrete::new(Concrete::from("hello"), Loc::Implicit);
+        let y = RangeConcrete::new(Concrete::from(""), Loc::Implicit);
+        let result = x.range_concat(&y).unwrap().maybe_concrete_value().unwrap();
+        assert_eq!(result.val, x.val);
+    }
+
+    #[test]
+    fn dyn_concrete_bytes() {
+        let g = DummyGraph::default();
+        let x = RangeDyn::from_concrete(
+            Concrete::from(vec![b'h', b'e', b'l', b'l', b'o']),
+            Loc::Implicit,
+        )
+        .unwrap();
+        let y = RangeConcrete::new(
+            Concrete::from(vec![b'w', b'o', b'r', b'l', b'd']),
+            Loc::Implicit,
+        );
+        let expected: Elem<_> = Elem::ConcreteDyn(
+            RangeDyn::from_concrete(
+                Concrete::from(vec![
+                    b'h', b'e', b'l', b'l', b'o', b'w', b'o', b'r', b'l', b'd',
+                ]),
+                Loc::Implicit,
+            )
+            .unwrap(),
+        );
+        let result = x.range_concat(&y).unwrap().maximize(&g).unwrap();
+        assert_eq!(result, expected);
+    }
+
+    #[test]
+    fn dyn_dyn_bytes() {
+        let g = DummyGraph::default();
+        let x = RangeDyn::from_concrete(
+            Concrete::from(vec![b'h', b'e', b'l', b'l', b'o']),
+            Loc::Implicit,
+        )
+        .unwrap();
+        let y = RangeDyn::from_concrete(
+            Concrete::from(vec![b'w', b'o', b'r', b'l', b'd']),
+            Loc::Implicit,
+        )
+        .unwrap();
+        let expected: Elem<_> = Elem::ConcreteDyn(
+            RangeDyn::from_concrete(
+                Concrete::from(vec![
+                    b'h', b'e', b'l', b'l', b'o', b'w', b'o', b'r', b'l', b'd',
+                ]),
+                Loc::Implicit,
+            )
+            .unwrap(),
+        );
+        let result = x.range_concat(&y).unwrap().maximize(&g).unwrap();
+        assert_eq!(result, expected);
+    }
+
+    #[test]
+    fn dyn_concrete_str() {
+        let g = DummyGraph::default();
+        let x = RangeDyn::from_concrete(Concrete::from("hello"), Loc::Implicit).unwrap();
+        let y = RangeConcrete::new(Concrete::from("world"), Loc::Implicit);
+        let expected: Elem<_> = Elem::ConcreteDyn(
+            RangeDyn::from_concrete(Concrete::from("helloworld"), Loc::Implicit).unwrap(),
+        );
+        let result = x.range_concat(&y).unwrap().maximize(&g).unwrap();
+        assert_eq!(result, expected);
+    }
+
+    #[test]
+    fn dyn_dyn_str() {
+        let g = DummyGraph::default();
+        let x = RangeDyn::from_concrete(Concrete::from("hello"), Loc::Implicit).unwrap();
+        let y = RangeDyn::from_concrete(Concrete::from("world"), Loc::Implicit).unwrap();
+        let expected: Elem<_> = Elem::ConcreteDyn(
+            RangeDyn::from_concrete(Concrete::from("helloworld"), Loc::Implicit).unwrap(),
+        );
+        let result = x.range_concat(&y).unwrap().maximize(&g).unwrap();
+        assert_eq!(result, expected);
     }
 }
